102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,13 +10,13 @@ int main(void)
 {
 	unsigned long int term1 = 1;
 	unsigned long int term2 = 2;
-	int next;
+	unsigned long int next;
 	int i;
 
-	printf("%d, ", term1);
+	printf("%lu, ", term1);
 	for (i = 1 ; i < 50 ; i++)
 	{
-		printf("%d, ", term2);
+		printf("%lu, ", term2);
 		next = term1 + term2;
 		term1 = term2;
 		term2 = next;
